Authentication

+++ frontend/src/components/WorkoutDetails.js
@@ -2,13 +2,21 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 import { toast } from 'react-toastify'; // Import toast for notifications
-
+import { useAuthContext } from '../hooks/useAuthContext';
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
-
+  const {user} = useAuthContext()
   const handleClick = async () => {
+    if(!user){
+
+      return
+    }
+    
     const response = await fetch('/api/workouts/' + workout._id, {
       method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     });
     const json = await response.json();
